requirements.txt
python-telegram-bot==20.8
yfinance
pandas
scikit-learn
xgboost
numpy
apscheduler